+++ HISTORY.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 * Migrate to pyrate-limiter v4
-* ⚠️ If you are using pyrate-limiter features directly, see its [release notes](https://github.com/vutran1710/PyrateLimiter/blob/master/CHANGELOG.md) for info on breaking changes.
+* ⚠️ Remove `max_delay`
+* ⚠️ If you are using pyrate-limiter features directly (via `Limiter` class or custom bucket classes), see its [release notes](https://github.com/vutran1710/PyrateLimiter/blob/master/CHANGELOG.md) for info on breaking changes.
 * ⚠️ Drop support for python 3.8 and 3.9 (required upstream)
 
 ## 0.8.0 (2026-01-03)